# swagger.yaml
swagger: '2.0'
info:
  version: '2.0'
  title: 'Content Management'
  description: 'REST APIs for document management'
  contact: {}
host: 'www.automate.io'
basePath: /v1
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /users/files:
    get:
      tags:
      - "Folders and Files"
      description: 'Get all files or folders of a context. A context can be root or folder'
      summary: 'get contents of folder or context'
      operationId: 'getContents'
      deprecated: false
      security:
      - automateio_auth:
        - "basic_user"
      produces:
        - application/json
      parameters:
        - name: authorization
          in: header
          type: 'string'
          required: true
        - name: userId
          in: query
          required: true
          type: 'string'
        - name: context
          in: query
          required: true
          type: 'string'
      responses:
        '200':
          description: ''
    put:
      tags:
      - "Folders and Files"
      description: 'Create file or folder'
      summary: 'create a content'
      operationId: 'createContent'
      deprecated: false
      security: []
      produces:
        - application/json
      parameters:
        - name: authorization
          in: header
          type: 'string'
          required: true
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/createFileOrFolder"
      responses:
        '200':
          description: ''
    delete:
      tags:
      - "Folders and Files"
      description: 'Delete file or folder'
      summary: 'remove a content'
      operationId: 'deleteContent'
      deprecated: false
      security: []
      produces:
        - application/json
      parameters:
        - name: authorization
          in: header
          type: 'string'
          required: true
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/deleteFileOrFolder"
      responses:
        '200':
          description: ''
    patch:
      tags:
      - "Folders and Files"
      description: 'Move file'
      summary: 'move a content'
      operationId: 'changeFileContext'
      deprecated: false
      security: []
      produces:
        - application/json
      parameters:
        - name: authorization
          in: header
          type: 'string'
          required: true
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/changeFileContext"
      responses:
        '200':
          description: ''
securityDefinitions:
  automateio_auth:
    type: "oauth2"
    authorizationUrl: "https://automate.io"
    flow: "implicit"
    scopes:
      basic_user: "all powers of a basic user"
definitions:
  changeFileContext:
    type: "object"
    properties:
      id:
        type: "string"
      context:
        type: "string"
  deleteFileOrFolder:
    type: "object"
    properties:
      id:
        type: "string"
  createFileOrFolder:
    type: "object"
    properties:
      name:
        type: "string"
      contentType: 
        type: "string"
        enum: [file, folder]
      format: 
        type: "string"
      fileId: 
        type: "string"
      context:
        type: "string"
...